+++ packages/dagu/src/Utils.js
@@ -2,9 +2,9 @@ import { composeNextState } from '@openfn/language-common';
 import {
   request as commonRequest,
   logResponse,
+  encodeFormBody,
 } from '@openfn/language-common/util';
 import nodepath from 'node:path';
-
 
 let access_token;
 
@@ -27,7 +27,7 @@ export const getAccessToken = async (configuration, headers) => {
   const { body } = await commonRequest('POST', '/account/login', {
     headers: {
       'content-type': 'application/json',
-      ...headers
+      ...headers,
     },
     baseUrl,
     parseAs: 'json',
@@ -41,24 +41,31 @@ export const getAccessToken = async (configuration, headers) => {
 };
 
 export const request = async (configuration = {}, method, path, options) => {
-  const { baseUrl  } = configuration;
+  const { baseUrl } = configuration;
 
-  if(!access_token)
-      access_token = await getAccessToken(configuration, options.headers)
+  if (!access_token)
+    access_token = await getAccessToken(configuration, options.headers);
 
-  const { query = {}, body = {} } = options;
+  const { query = {}, body = {}, contentType, headers = {} } = options;
+
+  let requestBody = body;
+  let requestHeaders = { ...headers };
+
+  if (contentType === 'form') {
+    requestBody = encodeFormBody(body);
+  } else {
+    requestHeaders['content-type'] = 'application/json';
+  }
+
+  requestHeaders['Authorization'] = `Bearer ${access_token}`;
 
   const opts = {
+    ...options,
     parseAs: 'json',
     baseUrl,
-    headers: {
-      'content-type': 'application/json',
-      Authorization: `Bearer ${access_token}`,
-      ...options.headers,
-    },
-    body,
+    headers: requestHeaders,
+    body: requestBody,
     query,
-    ...options,
   };
 
   const safePath = nodepath.join(path);