+++ packages/fhir-eswatini/src/profiles/SzAppointment.ts
@@ -149,5 +149,11 @@ export default function(props: Partial<Appointment_SzAppointment_Props>) {
         }
     }
 
+    resource.text = {
+      status: 'generated',
+      div: `<div xmlns=\"http://www.w3.org/1999/xhtml\">
+      <h2>${resource.resourceType}: ${resource.id || '(anon)'}</h2>
+</div>`,
+    };
     return resource;
 }