+++ packages/fhir-eswatini/src/profiles/SzLabSpecimen.ts
@@ -123,5 +123,11 @@ export default function(props: Partial<Specimen_SzLabSpecimen_Props>) {
         dt.ensureConceptText(resource.condition);
     }
 
+    resource.text = {
+      status: 'generated',
+      div: `<div xmlns=\"http://www.w3.org/1999/xhtml\">
+      <h2>${resource.resourceType}: ${resource.id || '(anon)'}</h2>
+</div>`,
+    };
     return resource;
 }