+++ packages/fhir-eswatini/src/profiles/SzPatient.ts
@@ -218,5 +218,11 @@ export default function(props: Partial<Patient_SzPatient_Props>) {
         }
     }
 
+    resource.text = {
+      status: 'generated',
+      div: `<div xmlns=\"http://www.w3.org/1999/xhtml\">
+      <h2>${resource.resourceType}: ${resource.id || '(anon)'}</h2>
+</div>`,
+    };
     return resource;
 }