+++ packages/fhir-eswatini/src/profiles/SzPractitioner.ts
@@ -84,5 +84,11 @@ export default function(props: Partial<Practitioner_SzPractitioner_Props>) {
         dt.ensureConceptText(resource.communication);
     }
 
+    resource.text = {
+      status: 'generated',
+      div: `<div xmlns=\"http://www.w3.org/1999/xhtml\">
+      <h2>${resource.resourceType}: ${resource.id || '(anon)'}</h2>
+</div>`,
+    };
     return resource;
 }