+++ packages/chatgpt/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @openfn/language-chatgpt
 
+## 2.0.10 - 30 March 2026
+
+### Patch Changes
+
+- Updated dependencies \[295655f]
+  - @openfn/language-common@3.3.0
+
 ## 2.0.9 - 24 February 2026
 
 ### Patch Changes