+++ packages/magpi/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @openfn/language-magpi
 
+## 1.2.10 - 30 March 2026
+
+### Patch Changes
+
+- Updated dependencies \[295655f]
+  - @openfn/language-common@3.3.0
+
 ## 1.2.9 - 24 February 2026
 
 ### Patch Changes