+++ packages/sunbird-rc/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @openfn/language-sunbird-rc
 
+## 1.0.6 - 30 March 2026
+
+### Patch Changes
+
+- Updated dependencies \[295655f]
+  - @openfn/language-common@3.3.0
+
 ## 1.0.5 - 24 February 2026
 
 ### Patch Changes