+++ packages/varo/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @openfn/language-varo
 
+## 2.1.9 - 30 March 2026
+
+### Patch Changes
+
+- Updated dependencies \[295655f]
+  - @openfn/language-common@3.3.0
+
 ## 2.1.8 - 24 February 2026
 
 ### Patch Changes