+++ packages/zata/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @openfn/language-zata
 
+## 1.0.9 - 30 March 2026
+
+### Patch Changes
+
+- Updated dependencies \[295655f]
+  - @openfn/language-common@3.3.0
+
 ## 1.0.8 - 24 February 2026
 
 ### Patch Changes